dia3/requirements.txt
django==3.2
djangorestframework
mysqlclient
djangorestframework-simplejwt